EDMT-383 refactor : 기존 signup 폴더에 있던 로그인/회원가입 스키마 auth 도메인 type 폴더로 이동 및 파일명 수정
